% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elp}
\alias{elp}
\title{Input lexicon ELP simulations}
\format{A data frame with 2,510 rows and 4 columns:
\describe{
  \item{\code{Word}}{a mono-syllabic mono-morphemic word}
  \item{\code{Pron}}{the pronunciation of the word in DISC format}
  \item{\code{Gestures}}{an approximation of the acoustic gestures required for the 
  pronunciation of the word through demi-syllables (see Klatt, 1979)}
  \item{\code{RTnaming}}{the naming latency for the word in the English Lexicon 
  Project (ELP, Balota et al. (2007)}
}}
\source{
Balota, D. A., Yap, M. J., Hutchison, K. A., Cortese, M. J., Kessler, B., 
Loftis, B., Neely, J. H., Nelson, D. L., Simpson, G. B., & Treiman, R. (2007). The 
English lexicon project. Behavior research methods, 39(3), 445-459.
}
\usage{
elp
}
\description{
The input lexicon used for the ELP simulations in Hendrix (2018).
}
\references{
Hendrix, P, Ramscar, M., & Baayen, R. H. (2019). NDRa: a single route model of 
response times in the reading aloud task based on discriminative learning. Manuscript.

Klatt, D. H. (1979). Speech perception: a model of acoustic-phonetic analysis and 
lexical access. Journal of Phonetics, 7, 279-312.
}
\keyword{datasets}
