% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestures.R
\name{gestures}
\alias{gestures}
\title{Define gestures}
\usage{
gestures(pron = "b8r")
}
\arguments{
\item{pron}{The pronunciation of a word or nonword in DISC notation}
}
\description{
Generate gestures for a word or nonword.
}
\examples{
# Load data for the ELP simulations in Hendrix (2018)
data(elp)

# Define gestures
elp$Gestures = sapply(elp$Phon, gestures)
}
