% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePronunciations.R
\name{simulatePronunciations}
\alias{simulatePronunciations}
\title{Generate simulated pronunciations}
\usage{
simulatePronunciations(lexicon = lex, weightsSem = weights_sem,
  weightsPhon = weights_phon, parallel = TRUE,
  numCores = detectCores(), verbose = TRUE)
}
\arguments{
\item{lexicon}{A dataframe with the colums "Word" and "Gestures". "Gestures" are 
demi-syllables (see Klatt, 1979) and can be generated using gestures().}

\item{weightsSem}{An orthography-to-semantics weight matrix with letter unigrams and 
bigrams as cues and words as outcomes. The default, "weights_sem" uses 
the weight matrix from Hendrix et al. (2018).}

\item{weightsPhon}{A phonology-to-semantic weight matrix with demi-syllables as 
cues and words as outcomes. The default, "weigths_phon" uses the weight matrix 
from Hendrix et al. (2018).}

\item{parallel}{Should computations be carried out in parallel? Defaults to TRUE.}

\item{numCores}{The number of cores to use for parallel computation. By default all 
available cores are used.}
}
\description{
Generate simulated pronunciation for a set of words or nonwords.
}
\examples{
# Load data for the ELP simulations in Hendrix (2018)
data(elp)

# Generate simulated pronunciations for a lexicon
elp$SimPron = simulatePronunciations(elp$Word)

}
\references{
Hendrix, P, Ramscar, M., & Baayen, R. H. (2019). NDRa: a single route model of 
response times in the reading aloud task based on discriminative learning. Manuscript.

Klatt, D. H. (1979). Speech perception: a model of acoustic-phonetic analysis and 
lexical access. Journal of Phonetics, 7, 279-312.
}
