% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weights_phon}
\alias{weights_phon}
\title{Phonology to semantics weight matrix}
\format{A matrix with 1,358 rows and 3,908 columns:
\describe{
  \item{\code{rows}}{demi-syllables (see Klatt, 1979)}
  \item{\code{columns}}{words}
}}
\source{
Hendrix, P, Ramscar, M., & Baayen, R. H. (2019). NDRa: a single route model of 
response times in the reading aloud task based on discriminative learning. Manuscript.
}
\usage{
weights_phon
}
\description{
Phonology to semantics weight matrix generated with version 0.2.18 of the ndl package.
}
\references{
Klatt, D. H. (1979). Speech perception: a model of acoustic-phonetic analysis and 
lexical access. Journal of Phonetics, 7, 279-312.
}
\keyword{datasets}
