% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weights_sem}
\alias{weights_sem}
\title{Orthography to semantics weight matrix}
\format{A matrix with 664 rows and 6,768 columns:
\describe{
  \item{\code{rows}}{letter unigrams and bigrams}
  \item{\code{columns}}{words}
}}
\source{
Hendrix, P, Ramscar, M., & Baayen, R. H. (2019). NDRa: a single route model of 
response times in the reading aloud task based on discriminative learning. Manuscript.
}
\usage{
weights_sem
}
\description{
Orthography to semantics weight matrix generated with version 0.2.18 of the ndl package.
}
\keyword{datasets}
